+++ test/object.test.ts
@@ -1,6 +1,8 @@
 import { keySelect } from '../src';
 
-test('keySelect', () => {
-  expect(keySelect('foo', { foo: 'bar' })).toBe('bar');
-  expect(keySelect(2, [10, 20, 25])).toBe(25);
+describe('Module "object"', () => {
+  test('keySelect', () => {
+    expect(keySelect('foo', { foo: 'bar' })).toBe('bar');
+    expect(keySelect(2, [10, 20, 25])).toBe(25);
+  });
 });